refactor(task): split task code into dedicated crates
