+++ crates/peekoo-agent-app/Cargo.toml
@@ -12,7 +12,8 @@ agent-client-protocol = "0.10"
 peekoo-agent = { path = "../peekoo-agent" }
 peekoo-app-settings = { path = "../peekoo-app-settings" }
 peekoo-agent-auth = { path = "../peekoo-agent-auth" }
-peekoo-productivity-domain = { path = "../peekoo-productivity-domain" }
+peekoo-task-domain = { path = "../peekoo-task-domain" }
+peekoo-task-app = { path = "../peekoo-task-app" }
 peekoo-pomodoro-app = { path = "../peekoo-pomodoro-app" }
 peekoo-persistence-sqlite = { path = "../persistence-sqlite" }
 peekoo-security = { path = "../security" }
