+++ crates/peekoo-agent-app/src/lib.rs
@@ -4,7 +4,6 @@ pub mod conversation;
 pub mod mcp_server;
 pub mod plugin;
 pub mod plugin_tool_impl;
-pub mod productivity;
 pub mod settings;
 pub mod task_parser;
 mod task_runtime_service;
@@ -17,9 +16,8 @@ pub use peekoo_app_settings::{AppSettingsService, SpriteInfo};
 pub use peekoo_notifications::PeekBadgeItem;
 pub use peekoo_plugin_store::{PluginSource, StorePluginDto};
 pub use peekoo_pomodoro_app::{PomodoroCycleDto, PomodoroSettingsInput, PomodoroStatusDto};
-pub use peekoo_productivity_domain::task::{TaskDto, TaskEventDto, TaskService};
+pub use peekoo_task_app::{SqliteTaskService, TaskDto, TaskEventDto, TaskService};
 pub use plugin::{PluginConfigFieldDto, PluginNotificationDto, PluginPanelDto, PluginSummaryDto};
-pub use productivity::ProductivityService;
 pub use settings::{
     AgentSettingsCatalogDto, AgentSettingsDto, AgentSettingsPatchDto, OauthCancelResponse,
     OauthStartResponse, OauthStatusRequest, OauthStatusResponse, ProviderAuthDto,
