+++ crates/peekoo-plugin-host/Cargo.toml
@@ -15,7 +15,8 @@ ed25519-dalek = { version = "2", features = ["rand_core"] }
 hex = "0.4"
 peekoo-notifications = { path = "../peekoo-notifications" }
 peekoo-agent-auth = { path = "../peekoo-agent-auth" }
-peekoo-productivity-domain = { path = "../peekoo-productivity-domain" }
+peekoo-task-domain = { path = "../peekoo-task-domain" }
+peekoo-task-app = { path = "../peekoo-task-app" }
 peekoo-paths = { path = "../peekoo-paths" }
 peekoo-scheduler = { path = "../peekoo-scheduler" }
 peekoo-security = { path = "../security" }
