+++ crates/peekoo-plugin-store/Cargo.toml
@@ -13,6 +13,7 @@ tracing = "0.1"
 ureq = { version = "2", features = ["json"] }
 
 peekoo-notifications = { path = "../peekoo-notifications" }
-peekoo-productivity-domain = { path = "../peekoo-productivity-domain" }
+peekoo-task-domain = { path = "../peekoo-task-domain" }
+peekoo-task-app = { path = "../peekoo-task-app" }
 peekoo-scheduler = { path = "../peekoo-scheduler" }
 rusqlite = { version = "0.38" }
