--- crates/peekoo-productivity-domain/AGENTS.md
@@ -1,20 +1,0 @@
-# AGENTS.md - peekoo-productivity-domain
-
-## Overview
-Domain models for productivity concerns in Peekoo AI.
-
-## Key Types
-- `Task` with `TaskPriority` and `TaskStatus`
-
-## Dependencies
-- `serde` for serialization
-- `thiserror` for domain errors
-
-## Testing
-```bash
-cargo test -p peekoo-productivity-domain
-```
-
-## Boundaries
-- Keep this crate pure domain logic.
-- No persistence, OAuth, transport, or UI concerns.
