+++ crates/peekoo-task-app/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "peekoo-task-app"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+chrono = { version = "0.4", default-features = false, features = ["clock", "std"] }
+peekoo-task-domain = { path = "../peekoo-task-domain" }
+rusqlite = { version = "0.38" }
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+thiserror = "1"
+tracing = "0.1"
+uuid = { version = "1", features = ["v4"] }
