+++ crates/peekoo-task-app/src/lib.rs
@@ -1,0 +1,7 @@
+mod dto;
+mod service;
+mod sqlite_task_service;
+
+pub use dto::{TaskDto, TaskEventDto};
+pub use service::{NoopTaskService, TaskService};
+pub use sqlite_task_service::SqliteTaskService;
