+++ crates/peekoo-task-domain/Cargo.toml
@@ -1,10 +1,10 @@
 [package]
-name = "peekoo-productivity-domain"
+name = "peekoo-task-domain"
 version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-thiserror = "1"
+chrono = { version = "0.4", features = ["serde"] }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
-chrono = { version = "0.4", features = ["serde"] }
+thiserror = "1"
