+++ crates/peekoo-task-domain/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod task;
+
+pub use task::{AgentWorkStatus, Task, TaskEvent, TaskEventType, TaskPriority, TaskStatus};
